Finished Frontend
Updated the bug issue of restaurants list page where it wouldn't display the restaurants in cards. - Issue was tied to the restaurant service file as the path was incorrect and the naming of the array object "restaurant" in restaurants.controller file was spelled incorrectly.

Added functions to view the restaurant and all of its reviews.

Added functions to add a review, delete/edit a review (only if you are that particular user).

Also created a login feature, but not so complicated with the auth.